TFG-REST refactor tests path
