+++ code/tests/mocks/mock_cie_repository.py
