+++ code/tests/mocks/mock_disease_repository.py
